+++ macros/src/attr/mod.rs
@@ -6,44 +6,16 @@ use quote::quote;
 pub use r#enum::*;
 pub use r#struct::*;
 use syn::{
+    parenthesized,
     parse::{Parse, ParseStream},
     punctuated::Punctuated,
     Error, Expr, Ident, Lit, Path, Result, Token, WherePredicate,
 };
 pub use variant::*;
-
 mod r#enum;
 mod field;
 mod r#struct;
 mod variant;
-
-/// Indicates whether the field is marked with `#[ts(optional)]`.
-/// `#[ts(optional)]` turns an `t: Option<T>` into `t?: T`, while
-/// `#[ts(optional = nullable)]` turns it into `t?: T | null`.
-#[derive(Default, Clone, Copy)]
-pub enum Optional {
-    Optional {
-        nullable: bool,
-    },
-
-    #[default]
-    NotOptional,
-}
-
-impl Optional {
-    pub fn or(self, other: Optional) -> Self {
-        match (self, other) {
-            (Self::NotOptional, Self::NotOptional) => Self::NotOptional,
-
-            (Self::Optional { nullable }, Self::NotOptional)
-            | (Self::NotOptional, Self::Optional { nullable }) => Self::Optional { nullable },
-
-            (Self::Optional { nullable: a }, Self::Optional { nullable: b }) => {
-                Self::Optional { nullable: a || b }
-            }
-        }
-    }
-}
 
 #[derive(Copy, Clone, Debug)]
 pub enum Inflection {
@@ -180,7 +152,7 @@ fn parse_concrete(input: ParseStream) -> Result<HashMap<syn::Ident, syn::Type>> 
     }
 
     let content;
-    syn::parenthesized!(content in input);
+    parenthesized!(content in input);
 
     Ok(
         Punctuated::<Concrete, Token![,]>::parse_terminated(&content)?
@@ -237,17 +209,22 @@ fn parse_bound(input: ParseStream) -> Result<Vec<WherePredicate>> {
     }
 }
 
-fn parse_optional(input: ParseStream) -> Result<Optional> {
-    let nullable = if input.peek(Token![=]) {
-        input.parse::<Token![=]>()?;
-        let span = input.span();
-        match Ident::parse(input)?.to_string().as_str() {
-            "nullable" => true,
-            _ => Err(Error::new(span, "expected 'nullable'"))?,
-        }
-    } else {
-        false
-    };
+fn parse_repr(input: ParseStream) -> Result<Repr> {
+    let content;
+    parenthesized!(content in input);
 
-    Ok(Optional::Optional { nullable })
+    content.parse::<Token![enum]>()?;
+
+    if content.is_empty() {
+        return Ok(Repr::Int);
+    }
+
+    content.parse::<Token![=]>()?;
+
+    let span = content.span();
+    let ident = content.parse::<Ident>()?;
+    match ident.to_string().as_str() {
+        "name" => Ok(Repr::Name),
+        _ => syn_err!(span; "expected `name`"),
+    }
 }