+++ ts-rs/tests/integration/optional_field.rs
@@ -141,6 +141,14 @@ struct NullableStruct {
 
     #[ts(as = "String")]
     h: Option<i32>,
+
+    // not nullable
+    #[ts(optional)]
+    i: Option<i32>,
+
+    // not optional
+    #[ts(optional = false)]
+    j: Option<i32>,
 }
 
 #[test]
@@ -148,7 +156,142 @@ fn struct_nullable() {
     assert_eq!(
         NullableStruct::inline(),
         format!(
-            "{{ a?: number | null, b?: number | null, c?: number | null, d: number, e?: number | null, f?: number | null, g: string, h: string, }}"
+            "{{ a?: number | null, b?: number | null, c?: number | null, d: number, e?: number | null, f?: number | null, g: string, h: string, i?: number, j: number | null, }}"
         )
     )
+}
+
+#[derive(Serialize, TS)]
+#[ts(export, export_to = "optional_field/")]
+struct OptionalInTuple(
+    Option<i32>,
+    #[ts(optional)] Option<i32>,
+    #[ts(optional = nullable)] Option<i32>,
+);
+
+#[test]
+fn in_tuple() {
+    assert_eq!(
+        OptionalInTuple::inline(),
+        format!("[number | null, (number)?, (number | null)?]")
+    );
+}
+
+#[derive(Serialize, TS)]
+#[ts(export, export_to = "optional_field/")]
+#[ts(optional_fields)]
+struct OptionalTuple(
+    i32,
+    #[ts(type = "string")] Option<i32>,
+    #[ts(as = "String")] Option<i32>,
+    Option<i32>,
+    #[ts(optional)] Option<i32>,
+    #[ts(optional = nullable)] Option<i32>,
+);
+
+#[test]
+fn tuple_optional() {
+    assert_eq!(
+        OptionalTuple::inline(),
+        "[number, string, string, (number)?, (number)?, (number | null)?]"
+    );
+}
+
+#[derive(Serialize, TS)]
+#[ts(export, export_to = "optional_field/")]
+#[ts(optional_fields = nullable)]
+struct NullableTuple(
+    i32,
+    #[ts(type = "string")] Option<i32>,
+    #[ts(as = "String")] Option<i32>,
+    Option<i32>,
+    #[ts(optional)] Option<i32>,
+    #[ts(optional = nullable)] Option<i32>,
+);
+
+#[test]
+fn tuple_nullable() {
+    assert_eq!(
+        NullableTuple::inline(),
+        "[number, string, string, (number | null)?, (number)?, (number | null)?]"
+    );
+}
+
+#[derive(Serialize, TS)]
+#[ts(export, export_to = "optional_field/", optional_fields)]
+enum OptionalFieldsEnum {
+    A { a: Option<i32> },
+    B { b: String, c: Option<bool> },
+}
+
+#[derive(Serialize, TS)]
+#[ts(export, export_to = "optional_field/", optional_fields = nullable)]
+enum OptionalFieldsEnumVariantOverride {
+    // Disable `nullable`
+    #[ts(optional_fields)]
+    A { a: Option<i32> },
+
+    // Disable `optional_fields`
+    #[ts(optional_fields = false)]
+    B { b: String, c: Option<bool> },
+}
+
+#[derive(Serialize, TS)]
+#[ts(export, export_to = "optional_field/", optional_fields)]
+enum OptionalFieldsEnumNotNullableVariantOverride {
+    // Disable `nullable`
+    #[ts(optional_fields = nullable)]
+    A { a: Option<i32> },
+
+    // Disable `optional_fields`
+    #[ts(optional_fields = false)]
+    B { b: String, c: Option<bool> },
+}
+
+#[derive(Serialize, TS)]
+#[ts(export, export_to = "optional_field/", optional_fields, tag = "type")]
+enum OptionalFieldsTaggedEnum {
+    A { a: Option<i32> },
+    B { b: String, c: Option<bool> },
+}
+
+#[derive(Serialize, TS)]
+#[ts(
+    export,
+    export_to = "optional_field/",
+    optional_fields,
+    tag = "type",
+    content = "data"
+)]
+enum OptionalFieldsExternallyTaggedEnum {
+    A { a: Option<i32> },
+    B { b: String, c: Option<bool> },
+}
+
+#[test]
+fn optional_fields_enum() {
+    assert_eq!(
+        OptionalFieldsEnum::inline(),
+        r#"{ "A": { a?: number, } } | { "B": { b: string, c?: boolean, } }"#
+    );
+
+    assert_eq!(
+        OptionalFieldsEnumVariantOverride::inline(),
+        r#"{ "A": { a?: number, } } | { "B": { b: string, c: boolean | null, } }"#
+    );
+
+    assert_eq!(
+        OptionalFieldsEnumNotNullableVariantOverride::inline(),
+        r#"{ "A": { a?: number | null, } } | { "B": { b: string, c: boolean | null, } }"#
+    );
+
+    assert_eq!(
+        OptionalFieldsTaggedEnum::inline(),
+        r#"{ "type": "A", a?: number, } | { "type": "B", b: string, c?: boolean, }"#
+    );
+
+    assert_eq!(
+        OptionalFieldsExternallyTaggedEnum::inline(),
+        r#"{ "type": "A", "data": { a?: number, } } | { "type": "B", "data": { b: string, c?: boolean, } }"#
+    );
 }